requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.1
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
Django==2.0.1
django-appconf==1.0.2
django-extensions==1.9.9
django-imagekit==4.0.2
django-model-utils==3.1.1
django-taggit==0.22.2
django-tinymce==2.7.0
easy-thumbnails==2.5
fmoo-audiotools==3.1b1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
mysqlclient==1.3.12
parso==0.1.1
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
pilkit==2.0
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pymediainfo==2.2.0
python-oembed==0.2.4
python-resize-image==1.1.11
python-slugify==1.2.4
pytz==2017.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
sorl-thumbnail==12.4.1
traitlets==4.3.2
typing==3.6.2
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7